fix: remove collection tasks for removed device
ZEN-34704
