+++ Products/ZenCollector/scheduler/__init__.py
@@ -1,0 +1,15 @@
+##############################################################################
+#
+# Copyright (C) Zenoss, Inc. 2024 all rights reserved.
+#
+# This content is made available according to terms specified in
+# License.zenoss under the directory where your Zenoss product is installed.
+#
+##############################################################################
+
+from __future__ import absolute_import
+
+from .scheduler import Scheduler, TaskScheduler
+
+
+__all__ = ("Scheduler", "TaskScheduler")
